cyber-war-room/requirements.txt
kafka-python
pytest
